LPS-147487 move HttpClient to com.liferay.site.initializer.testray.extra.java.function.http
